+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- [#532](https://github.com/pybop-team/PyBOP/issues/532) - Allows the user to update linked parameters during design optimisation.
 - [#529](https://github.com/pybop-team/PyBOP/issues/529) - Adds `GravimetricPowerDensity` and `VolumetricPowerDensity` costs, along with the mathjax extension for Sphinx.
 
 ## Optimisations